Reorganizing modules
Mutation and selection classes were moved to separate submodule.
Imports were moved to top level init, in order to simplify access.
