+++ pystrand/operators/__init__.py
@@ -1,0 +1,4 @@
+"""Pystrand genetic operators submodule.
+
+Contains genetic operators used by optimizers.
+"""
