+++ pystrand/operators/selections.py
