+++ tests/test_genotype_manipulation.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 from pystrand.genotypes import Genotype
-from pystrand.mutations import PointMutation
+from pystrand.operators.mutations import PointMutation
 
 
 class Test_genotype_manipulation(unittest.TestCase):
